Refactor: Merge resource_monitor and system_status into system_monitor.py (#185)
