--- engine/resource_monitor.py
@@ -1,23 +1,0 @@
-import psutil
-import shutil
-
-
-def get_resource_usage():
-    """Return system resource usage."""
-    try:
-        cpu = psutil.cpu_percent(interval=1)
-        memory = psutil.virtual_memory()
-        disk = shutil.disk_usage("/")
-
-        return {
-            "cpu": cpu,
-            "ram_percent": memory.percent,
-            "ram_used_gb": round(memory.used / (1024**3), 2),
-            "ram_total_gb": round(memory.total / (1024**3), 2),
-            "disk_percent": round((disk.used / disk.total) * 100, 2),
-            "disk_used_gb": round(disk.used / (1024**3), 2),
-            "disk_total_gb": round(disk.total / (1024**3), 2),
-        }
-
-    except Exception:
-        return None
